Moved common scss folder and corrected styles
