+++ app/screens/Library/components/Book/styles.js
@@ -2,11 +2,13 @@ import { StyleSheet } from 'react-native';
 
 import { elevationShadowStyle } from '../../../../../utils';
 
-import { white, darkGray, softGray } from '../../../../scss/colors';
-import { mediumSize } from '../../../../scss/fonts';
+import { white, darkGray, softGray } from '../../../../../scss/colors';
+import { mediumSize } from '../../../../../scss/fonts';
+
 
 export const styles = StyleSheet.create({
   bookContainer: {
+    ...elevationShadowStyle(5),
     backgroundColor: white,
     borderRadius: 5,
     height: 90,
@@ -16,7 +18,6 @@ export const styles = StyleSheet.create({
     marginHorizontal: 20,
     marginVertical: 10
   },
-  shadowBox: elevationShadowStyle(5),
   bookImage: {
     height: 60,
     width: 40,
