+++ app/screens/Library/styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 
-import { mainBackgroundColor } from '../../scss/colors';
+import { mainBackgroundColor } from '../../../scss/colors';
 
 export const styles = StyleSheet.create({
   container: {
