+++ scss/colors.js
@@ -1,4 +1,5 @@
 export const mainBackgroundColor = '#F2F3F4';
-export const white = '#FFFFFF';
+export const white = '#FFF';
+export const black = '#000'
 export const darkGray = '#4d4d4d';
 export const softGray = '#999999';
