+++ utils/index.js
@@ -1,7 +1,9 @@
+import { black } from '../scss/colors';
+
 export function elevationShadowStyle(elevation) {
   return {
     elevation,
-    shadowColor: 'black',
+    shadowColor: black,
     shadowOffset: { width: 0, height: 0.5 * elevation },
     shadowOpacity: 0.3,
     shadowRadius: 0.8 * elevation
